requirements.txt
Django==4.2.3
python-decouple==3.8
djangorestframework==3.14.0
dj-rest-auth==4.0.1
djangorestframework-simplejwt==5.2.2
